Change nginx configuration to be more reliable.
The nginx configuration generated by the devpi-server utility is pretty
opinionated (fragile?) when it comes to the IP configuration.  It was
assuming that the server is registered in DNS which is often not the
case.  A little adjustment to the nginx configuration fixes that.

If you curious, a PR against to devpi-server configuration template
has been issued to fix this problem at the source.

https://bitbucket.org/hpk42/devpi/pull-request/186/