tests/server/nginx_installation: Fix HTTP program usage.
Depending on how the test servers are constructed they will have
either wget or curl installed.  When I switched to using kitchen-docker
wget was no longer available under Ubuntu.  The solution was to add
a helper function that used either wget or curl depending on which
one is actually available.